requirements.txt
amqp==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_full_version <= "3.11.2"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
billiard==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
boto3==1.28.48 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.31.48 ; python_version >= "3.10" and python_version < "4.0"
celery==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
django-cacheops==6.2 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
django-dry-rest-permissions==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-filter==22.1 ; python_version >= "3.10" and python_version < "4.0"
django-storages==1.14 ; python_version >= "3.10" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.25.1 ; python_version >= "3.10" and python_version < "4.0"
funcy==1.18 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.19.0 ; python_version >= "3.10" and python_version < "4.0"
kombu==5.3.2 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.4.4 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.7 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
razorpay==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
redis==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.30.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.10.3 ; python_version >= "3.10" and python_version < "4.0"
s3transfer==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
